chown + tests
